Update and rename basicof stack.c to stack.c
